requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
dj-database-url==1.3.0
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
python-dotenv==0.20.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
zipp==3.17.0
